refactor: 불필요한 코드제거후 Unique제약조건위배시 에러핸들링 추가

+++ src/main/java/com/example/inflearn/ui/post/dto/response/PostResponse.java
@@ -12,9 +12,9 @@ public record PostResponse(
         String nickname,
         String title,
         String contents,
-        int viewCount,
-        Long likeCount,
-        Long commentCount,
+        long viewCount,
+        long likeCount,
+        long commentCount,
         Set<String> hashtags,
         LocalDateTime createdAt,
         PostStatus postStatus
